fuzz: make it possible to fuzz wazevo
