+++ internal/integration_test/fuzz/README.md
@@ -11,9 +11,9 @@ Fuzzing infrastructure for wazero engines via [wasm-tools](https://github.com/by
 
 Currently, we have the following fuzzing targets:
 
-- `basic`: compares the results from the compiler and interpreter engines, and see if there's a diff in them.
-- `memory_no_diff`: same as `basic` except that in addition to the results, it also compares the entire memory buffer between engines to ensure the consistency around memory access.
-  Therefore, this takes much longer than `basic`.
+- `no_diff`: compares the results from the compiler and interpreter engines, and see if there's a diff in them.
+- `memory_no_diff`: same as `no_diff` except that in addition to the results, it also compares the entire memory buffer between engines to ensure the consistency around memory access.
+  Therefore, this takes much longer than `no_diff`.
 - `validation`: try compiling maybe-invalid Wasm module binaries. This is to ensure that our validation phase works correctly as well as the engines do not panic during compilation.
 
 
@@ -22,9 +22,6 @@ To run the fuzzer on a target, execute the following command:
 ```
 # Running on the host archictecture.
 cargo fuzz run <target>
-
-# Running on the specified architecture which is handy when developping on M1 Mac.
-cargo fuzz run <target>-x86_64-apple-darwin
 ```
 
 where you replace `<target>` with one of the targets described above.
@@ -32,17 +29,17 @@ where you replace `<target>` with one of the targets described above.
 See `cargo fuzz run --help` for the options. Especially, the following flags are useful:
 
 - `-jobs=N`: `cargo fuzz run` by default only spawns one worker, so this flag helps do the parallel fuzzing.
-  - usage: `cargo fuzz run basic -- -jobs=5` will run 5 parallel workers to run fuzzing jobs.
+  - usage: `cargo fuzz run no_diff -- -jobs=5` will run 5 parallel workers to run fuzzing jobs.
 - `-max_total_time`: the maximum total time in seconds to run the fuzzer.
-  - usage: `cargo fuzz run basic -- -max_total_time=100` will run fuzzing for 100 seconds.
+  - usage: `cargo fuzz run no_diff -- -max_total_time=100` will run fuzzing for 100 seconds.
 - `-timeout` sets the timeout seconds _per fuzzing run_, not the entire job.
 - `-rss_limit_mb` sets the memory usage limit which is 2GB by default. Usually 2GB is not enough for some large Wasm binary.
 
 #### Example commands
 
 ```
-# Running the `basic` target with 15 concurrent jobs with total runnig time with 2hrs and 8GB memory limit.
-$ cargo fuzz run basic -- -rss_limit_mb=8192 -max_len=5000000 -max_total_time=7200 -jobs=15
+# Running the `no_diff` target with 15 concurrent jobs with total runnig time with 2hrs and 8GB memory limit.
+$ cargo fuzz run no_diff -- -rss_limit_mb=8192 -max_len=5000000 -max_total_time=7200 -jobs=15
 
 # Running the `memory_no_diff` target with 15 concurrent jobs with timeout 2hrs and setting timeout per fuzz case to 30s.
 $ cargo fuzz run memory_no_diff -- -timeout=30 -max_total_time=7200 -jobs=15
@@ -73,7 +70,7 @@ Also, in the bottom of the output, you can find the message as
 
 Minimize test case with:
 
-        cargo fuzz tmin basic fuzz/artifacts/basic/crash-d2c1f5307fde6f057454606bcc21d5653be9be8d
+        cargo fuzz tmin no_diff fuzz/artifacts/no_diff/crash-d2c1f5307fde6f057454606bcc21d5653be9be8d
 
 ────────────────────────────────────────────────────────────────────────────────
 ```
