+++ internal/integration_test/fuzz/wazerolib/lib.go
@@ -7,6 +7,10 @@ import (
 	"fmt"
 	"os"
 	"path"
+
+	"github.com/tetratelabs/wazero"
+	"github.com/tetratelabs/wazero/api"
+	"github.com/tetratelabs/wazero/internal/engine/wazevo"
 )
 
 func main() {}
@@ -68,4 +72,14 @@ Failed Wasm binary has been written to %s
 To reproduce the failure, execute: WASM_BINARY_PATH=%s go test -run=%s ./wazerolib/...
 `, hex.EncodeToString(bin), binaryPath, binaryPath, reproduceTestName)
 	}
+}
+
+// This returns a wazevo.RuntimeConfigure whose compiler is either wazevo or the default.
+func newCompilerConfig() wazero.RuntimeConfig {
+	c := wazero.NewRuntimeConfigCompiler()
+	if os.Getenv("WAZERO_FUZZ_WAZEVO") != "" {
+		c = c.WithCoreFeatures(api.CoreFeaturesV1) // Currently only V1 is supported.
+		wazevo.ConfigureWazevo(c)
+	}
+	return c
 }
